requirements.txt
appium-python-client
pytest~=8.4.2
allure-pytest
python-dotenv~=1.1.1
selene>=2.0.0rc4
allure-python-commons~=2.15.0
selenium
requests